Trait compgeo::line::DistanceToPoint

pub trait DistanceToPoint {
    fn distance_to_point(&self, point: &Point2<f32>) -> f32;
    fn distance_to_point_squared(&self, point: &Point2<f32>) -> f32;
}

Types which implement this trait can compute their distance from an arbitrary point.

Required methods

fn distance_to_point(&self, point: &Point2<f32>) -> f32

Compute the L2 Norm distance from this object to an arbitrary point.

fn distance_to_point_squared(&self, point: &Point2<f32>) -> f32

Compute the squared L2 Norm distance from this object to an arbritrary point.